Initial commit of Poission and WaveEqn
Poisson contains a working program on a hypermesh.  We need to update
this program with a cooler mesh (probably our pipe mesh)

WaveEqn contains all the build files.  I did not commit the source file
because it is not finished.  The source file can be found in my development
branch.