Working version of WaveEqn
The first rev of the WaveEqn program is completed and
commented.  The program computes the wave equation on a fixed
mesh over 5 seconds, with a step size of 1/64 second.

Future improvements could include adaptivly changing the mesh
based on the requirements of the solution, and making it
easier to change the time step, total time, and resolution.