Added a Hypermesh generated file of a test cylinder
Uploaded a test mesh in abaqus file format of a test
cylinder generated in Hypermesh.  The abaqus file format
was chosen because I didn't know what else to export it
as.  Abaqus appears to have a fairly parseable file
format, so it's possible that we can make a parser to
go from abaqus to gmesh files.